feat(falcon-script-watch): fixed watch for esm build
