--- packages/falcon-dev-tools/falcon-scripts/src/watch-build.js
@@ -1,19 +1,0 @@
-const spawn = require('cross-spawn');
-
-module.exports = () => {
-  console.log('building esm...');
-
-  const babelConfigPath = require.resolve('./babel/babel.config');
-
-  const result = spawn.sync(
-    `babel`,
-    ['src', '-d', 'dist', '-x', '.ts,.tsx', '-s', '--watch', '--config-file', babelConfigPath],
-    {
-      stdio: 'inherit'
-    }
-  );
-
-  if (result.status !== 0) {
-    throw result;
-  }
-};
